Move Pathfinding for roads out into own file
